Battle classes and some character
DungeonDefenders is the tester
Encounter class holds all the battle methods
the rest is the class that i created to go with the class.
I still am having problems to get them to work together. the character
classes are quick makes and can be changed. I added them to my upload
folder so you can use them as references.